requirements.txt
pytest==8.1.1
selenium==4.18.1
pytest-xdist==3.5.0
allure-pytest==2.13.5
requests==2.31.0
faker==24.4.0
pyyaml==6.0.1
tenacity==8.2.3
